+++ src/test/java/chess/domain/board/BoardTest.java
@@ -3,8 +3,8 @@ package chess.domain.board;
 import chess.TestPiecesFactory;
 import chess.domain.Color;
 import chess.domain.Position;
-import chess.domain.board.maker.PiecesFactory;
 import chess.domain.piece.*;
+import chess.domain.piecesfactory.PiecesFactory;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
