+++ src/test/java/chess/domain/board/maker/StartingPiecesFactoryTest.java
@@ -5,6 +5,7 @@ import chess.domain.File;
 import chess.domain.Position;
 import chess.domain.Rank;
 import chess.domain.piece.*;
+import chess.domain.piecesfactory.StartingPiecesFactory;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
